requirements.txt
flask
flask-restful
requests
flask-script
flask-cors
flask-sqlalchemy
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
celery[redis]
celery
bcrypt
flask-redis
Flask-JWT-Extended
pylint
gunicorn
pytest
pika
autopep8
pylint
python-dotenv
pyserial
